fixed tweet skipping bug
changed from Find to FindOne. removed debugging console outputs.

Find will return a string of outputs that match that id. The test of `if(exists)` will always return true due to it being a string. FindOne will only return one element hence the if statement will be false if that element is not found.